move terraform package from internal to pkg
