+++ internal/render/moved_block_renderer.go
@@ -1,4 +1,4 @@
-package terraform
+package render
 
 import (
 	"bytes"
