+++ internal/render/unified_diff_renderer.go
@@ -1,4 +1,4 @@
-package terraform
+package render
 
 import (
 	"bytes"
